Xcode: Fix bug generating code coverage data files (.gcda)
http://stackoverflow.com/questions/19136767/generate-gcda-files-with-xcode5-ios7-simulator-and-xctest
Workaround for XCode 5 bug where __gcov_flush is not called properly 
when Test Coverage flags are set, therefore not generating the
corresponding gcov data files.

Also, all the .m files of the "app target" need to be added to the 
Build Phase of the "test target" for their gcda files to be generated.